stable version with plenty of new features, removing solidity builded jsons

+++ ebsi-access-control/src/contracts_connections/Users.ts
@@ -2,7 +2,7 @@ import { store } from "../store/store";
 import { setCreatedUser, setCurrentUserEbsiDID, setDeletedUser, setEbsiDIDs, setUpdatedUser, setUsers } from "../store/accessControlListStore";
 import useAlert from "../hooks/useAlert";
 import { contract } from "./ContractsConnections";
-import { EBSI_DID_RESOLVED } from "../constants/Constants";
+import { EBSI_DID_RESOLVED, mockRole } from "../constants/Constants";
 import { _arrayBufferToBase64, addNewLines, fromBytes32ToString, fromStringToBytes32, translateTimeFromSol } from "../utilities/utilities";
 import { util } from "@cef-ebsi/key-did-resolver";
 import ResourceRole from "../models/ResourceRole";
@@ -48,7 +48,18 @@ const transformResourceRoles = (resourceRoles: ResourceRole[]) => {
 };
 
 const extractResourceNamesPayload = (resourceRoles: ResourceRole[]) =>
-    resourceRoles?.map((resourceRole: ResourceRole) => fromStringToBytes32(resourceRole.resourceName)) || [];
+    Array.from(new Set(resourceRoles?.map((resourceRole: ResourceRole) => fromStringToBytes32(resourceRole.resourceName)))) || [];
+
+const addNewResourcesOnly = (resources: string[], encodedResourceRoles: any[]) => {
+    resources = resources.map(resource => fromStringToBytes32(resource));
+    encodedResourceRoles.forEach((encodedResourceRole => {
+        if (resources.indexOf(encodedResourceRole.resourceName) < 0) {
+            resources.push(encodedResourceRole.resourceName);
+        }
+    }));
+
+    return resources;
+};
 
 export const convertPrivateToPEM = (privateKey: any) => {
     const b64 = addNewLines(_arrayBufferToBase64(privateKey));
@@ -56,6 +67,7 @@ export const convertPrivateToPEM = (privateKey: any) => {
 
     return pem;
 };
+
 
 const transformUser = (item: any): User => ({
     ebsiDID: item.ebsiDID,
@@ -65,54 +77,57 @@ const transformUser = (item: any): User => ({
     lastUpdate: translateTimeFromSol(item.lastUpdate).getTime()
 });
 
-const transformUsers = (response: any): User[] => {
-    return response?.map((item: any) => transformUser(item));
-};
+const transformUsers = (response: any): User[] => response?.map((item: any) => transformUser(item));
 
 export const listenForUsersEvents = () => {
     try {
         if (contract) {
             contract.events.UserUpdated()
                 .on('data', (event: any) => {
-                    const { users, ebsiDIDs } = store.getState().accessControlList || {};
+                    const {
+                        users,
+                        ebsiDIDs,
+                        createdUser,
+                        deletedUser,
+                        updatedUser
+                    } = store.getState().accessControlList || {};
+
                     switch (event?.returnValues?.eventType) {
                         case "creation": {
                             const userEbsiDID = event?.returnValues?.ebsiDID;
-                            if (!users.some((user: User) => user.ebsiDID === userEbsiDID)) {
-                                store.dispatch(setCreatedUser({ status: "success", ebsiDID: userEbsiDID }));
-                                store.dispatch(setEbsiDIDs([...ebsiDIDs, userEbsiDID]));
-                                requestUsers();
+                            if (!createdUser?.ebsiDID) {
+                                if (!users.some((user: User) => user.ebsiDID === userEbsiDID)) {
+                                    store.dispatch(setCreatedUser({ status: "success", ebsiDID: userEbsiDID }));
+                                    store.dispatch(setEbsiDIDs([...ebsiDIDs, userEbsiDID]));
+                                    requestUsers();
+                                }
+
                             }
 
                             break;
                         }
                         case 'deletion': {
-                            debugger;
-                            const userEbsiDID = event?.returnValues?.ebsiDID;
-                            store.dispatch(setDeletedUser({ status: "success", ebsiDID: userEbsiDID }));
-                            store.dispatch(setUsers(users.filter(item => item.ebsiDID !== userEbsiDID)));
-                            store.dispatch(setEbsiDIDs(ebsiDIDs.filter(item => item !== userEbsiDID)));
+                            if (!deletedUser?.ebsiDID) {
+                                const userEbsiDID = event?.returnValues?.ebsiDID;
+                                store.dispatch(setDeletedUser({ status: "success", ebsiDID: userEbsiDID }));
+                                store.dispatch(setUsers(users.filter(item => item.ebsiDID !== userEbsiDID)));
+                                store.dispatch(setEbsiDIDs(ebsiDIDs.filter(item => item !== userEbsiDID)));
+                                requestUsers();
+                            }
 
                             break;
                         }
                         case 'updated-resource-added':
-                        case 'updated-resource-removed': 
-                        {
-                            debugger;
-                            const userEbsiDID = event?.returnValues?.ebsiDID;
-                            store.dispatch(setUpdatedUser({ status: "success", ebsiDID: userEbsiDID }));
-                            requestUsers();
+                        case 'updated-resource-removed':
+                            {
+                                if (!updatedUser?.ebsiDID) {
+                                    const userEbsiDID = event?.returnValues?.ebsiDID;
+                                    store.dispatch(setUpdatedUser({ status: "success", ebsiDID: userEbsiDID }));
+                                    requestUsers();
+                                }
 
-                            break;
-                        }
-                        case 'updated-resource-removed': {
-                            debugger;
-                            const userEbsiDID = event?.returnValues?.ebsiDID;
-                            store.dispatch(setUpdatedUser({ status: "success", ebsiDID: userEbsiDID }));
-                            requestUsers();
-
-                            break;
-                        }
+                                break;
+                            }
                         default: {
                             requestUsers();
                         }
@@ -128,13 +143,12 @@ export const requestUsersEbsiDIDsArray = async () => {
     try {
         if (contract) {
             const response = await contract.methods.getAllEbsiDIDs().call();
-            debugger;
-            setEbsiDIDs(response);
+            store.dispatch(setEbsiDIDs(response));
         }
     } catch (error) {
         console.error(error);
         useAlert("Couldn't retrieve list of users", "red");
-        setEbsiDIDs([]);
+        store.dispatch(setEbsiDIDs([]));
     }
 };
 
@@ -144,22 +158,23 @@ export const requestUsers = async () => {
             const response = await contract?.methods.getAllUsers().call();
             const newUsers: User[] = transformUsers(response);
             store.dispatch(setUsers(newUsers));
-            return;
         }
     } catch (error) {
         console.log(error);
         useAlert("Something went wrong!", "red");
+        store.dispatch(setUsers([]));
     }
-    store.dispatch(setUsers([]));
 };
 
 export const requestUser = async (ebsiDID: string) => {
     try {
         if (contract) {
-            const response = await contract?.methods.getUser(ebsiDID).call();
-            const newUsers: User[] = transformUsers(response);
-            store.dispatch(setUsers(newUsers));
-            return;
+            const response = await contract?.methods.getUser(`did:key:${ebsiDID}`).call();
+            if (response?.ebsiDID) {
+                const newUsers: User[] = transformUsers(response);
+                store.dispatch(setUsers(newUsers));
+            }
+            useAlert("No user found for that ebsiDID");
         }
     } catch (error) {
         console.error(error);
@@ -218,7 +233,7 @@ export const createUser = async (data: any) => {
             const response = await contract.methods.createUser(newEbsiDID, resourceNames).send({ from: connectedAccount });
             useAlert("User created", "green");
             await encodedResourceRoles.forEach(async (encodedResourceRole) => {
-                await contract.methods.assignResourceToUser(newEbsiDID, encodedResourceRole.resourceName, encodedResourceRole.role).send({ from: connectedAccount })
+                await contract.methods.updateUserResources(newEbsiDID, resourceNames, encodedResourceRole.resourceName, encodedResourceRole.role, "creation").send({ from: connectedAccount })
             });
 
             return response;
@@ -227,7 +242,6 @@ export const createUser = async (data: any) => {
         console.error(error);
         const deniedOp: boolean = JSON.stringify(error).indexOf('User denied transaction signature') > -1;
         store.dispatch(setCreatedUser({ status: "error", ebsiDID: '' }));
-        debugger;
         const msg: string = deniedOp ? "Failed to create user: User rejected transaction" : "Something went wrong";
         useAlert(msg, "red");
         return {};
@@ -241,7 +255,7 @@ export const requestDeleteUser = async (user: User) => {
 
     try {
         if (contract) {
-            const response = await contract.method.removeUser(ebsiDID, newUsersArray).send({ fromAccount: connectedAccount });
+            const response = await contract.methods.removeUser(ebsiDID, newUsersArray).send({ from: connectedAccount });
             if (response) {
                 useAlert(`User: ${ebsiDID} correctly deleted!`, "green");
             }
@@ -251,6 +265,40 @@ export const requestDeleteUser = async (user: User) => {
         const deniedOp: boolean = JSON.stringify(error).indexOf('User denied transaction signature') > -1;
         store.dispatch(setDeletedUser({ status: "error", ebsiDID }));
         const msg: string = deniedOp ? "Failed to delete user: User rejected transaction" : "Something went wrong";
+        useAlert(msg, "red");
+        return;
+    }
+};
+
+export const updateUser = async (user: User | null, resourceRoles: ResourceRole[] = [], resourcesFromForm: string[], action: "creation" | "deletion" | null) => {
+    const { connectedAccount } = store.getState().accessControlList || {};
+    const { ebsiDID, resources = [] } = user || {};
+    const encodedResourceRoles = transformResourceRoles(resourceRoles);
+    const newResources = addNewResourcesOnly(resources, encodedResourceRoles) || [];
+
+    try {
+        if (contract && user && action) {
+            if (encodedResourceRoles?.length > 0 && action === 'creation') {
+                await encodedResourceRoles.forEach(async (encodedResourceRole) => {
+                    await contract.methods.updateUserResources(ebsiDID, newResources, encodedResourceRole.resourceName, encodedResourceRole.role, action).send({ from: connectedAccount });
+                });
+            } else if (resourcesFromForm.length > 0 || action === 'deletion') {
+                // Here this will remove every selected resource from the new resource array, this is better than filtering for each one before launching the call to solidity, it avoids any problems related to asynchronous calls
+                const encodedResources: string[] = resources
+                    ?.filter(resource => !resourcesFromForm.some(resourceFromForm => resourceFromForm === resource))
+                    ?.map(resource => fromStringToBytes32(resource));
+
+                await resourcesFromForm.forEach(async (resource) => {
+                    const resourceToDelete: string = fromStringToBytes32(resource);
+                    await contract.methods?.updateUserResources(ebsiDID, encodedResources, resourceToDelete, mockRole, 'deletion').send({ from: connectedAccount });
+                });
+            }
+        }
+    } catch (error) {
+        console.error(error);
+        const deniedOp: boolean = JSON.stringify(error).indexOf('User denied transaction signature') > -1;
+        store.dispatch(setUpdatedUser({ status: "error", ebsiDID }));
+        const msg: string = deniedOp ? "Failed to update user: User rejected transaction" : "Something went wrong";
         useAlert(msg, "red");
         return;
     }
